Solve some of the conflict
