+++ src/componenets/auth/Registration.js
@@ -55,9 +55,15 @@ class Registration extends Component {
     const { email, password, passwordConfirmation } = this.state;
 
     return (
-      <div>
-        <form onSubmit={this.handleSubmit}>
+      <div className="add-house-section"> 
+      <div className="backdrop column flex-center">
+      <h1>Sign Up</h1>
+          <p>
+            Hello there! Sign up and start managing your system
+          </p>
+        <form onSubmit={this.handleSubmit} className="column">
           <input
+            className="input-text"
             type="email"
             name="email"
             placeholder="Email"
@@ -67,6 +73,7 @@ class Registration extends Component {
           />
 
           <input
+            className="input-text"
             type="password"
             name="password"
             placeholder="Password"
@@ -76,6 +83,7 @@ class Registration extends Component {
           />
 
           <input
+            className="input-text"
             type="password"
             name="passwordConfirmation"
             placeholder="Password confirmation"
@@ -84,8 +92,9 @@ class Registration extends Component {
             required
           />
 
-          <button type="submit">Register</button>
+          <button type="submit" className='btn'>Register</button>
         </form>
+      </div>
       </div>
     );
   }
