--- src/reducers/index.js
@@ -1,7 +1,0 @@
-import { combineReducers } from 'redux';
-
-const rootReducer = combineReducers({
-  state: (state = {}) => state,
-});
-
-export default rootReducer;
